+++ lib/src/models/documents/nodes/container.dart
@@ -14,7 +14,7 @@ import 'node.dart';
 ///
 /// Most of the operation handling logic is implemented by [Line]
 /// and [QuillText].
-abstract class Container<T extends Node?> extends Node {
+abstract base class Container<T extends Node?> extends Node {
   final LinkedList<Node> _children = LinkedList<Node>();
 
   /// List of children.