+++ lib/src/widgets/keyboard_listener.dart
@@ -44,8 +44,9 @@ class QuillKeyboardListenerState extends State<QuillKeyboardListener> {
   final QuillPressedKeys _pressedKeys = QuillPressedKeys();
 
   bool _keyEvent(KeyEvent event) {
-    _pressedKeys
-        ._updatePressedKeys(HardwareKeyboard.instance.logicalKeysPressed);
+    _pressedKeys._updatePressedKeys(
+      HardwareKeyboard.instance.logicalKeysPressed,
+    );
     return false;
   }